Bugfix/ability scores (#22)
My understanding of ability scores was incorrect. The value cap isn't additive. This has cascading effects for a lot of passives that grant ability bonuses. I don't want to maintain them all to remove their explicit cap. So, instead, I add notes for feats and passives that do this so that players can be informed beforehand.