Fix missing passives in the passive selector
Some passives belong to a category (in particular, WarlockInvocations), that rely on the progression tables to determine which of many lists to present choices to during level up. However, the feat can only select one id and that may be low level. Now, when looking at passives, if they specify a category, gather any other passive selection instances from the progression tables that specify the same category, and merge the sets.

Note: progression table entries will only be included if the level is less than or equal to the max level for the mod.